+++ src/probnum/randprocs/markov/continuous/integrator/_integrator.py
@@ -159,464 +159,3 @@ class Integrator:
         """
         projmat = Integrator(ordint, spatialdim)._derivwise2coordwise_projmat
         return projmat @ state
-
-
-class IBM(Integrator, _sde.LTISDE):
-    """Integrated Brownian motion in :math:`d` dimensions."""
-
-    def __init__(
-        self,
-        ordint,
-        spatialdim,
-        forward_implementation="classic",
-        backward_implementation="classic",
-    ):
-        # initialise BOTH superclasses' inits.
-        # I don't like it either, but it does the job.
-        Integrator.__init__(self, ordint=ordint, spatialdim=spatialdim)
-        _sde.LTISDE.__init__(
-            self,
-            driftmat=self._driftmat,
-            forcevec=self._forcevec,
-            dispmat=self._dispmat,
-            forward_implementation=forward_implementation,
-            backward_implementation=backward_implementation,
-        )
-
-    @cached_property
-    def _driftmat(self):
-        driftmat_1d = np.diag(np.ones(self.ordint), 1)
-        return np.kron(np.eye(self.spatialdim), driftmat_1d)
-
-    @cached_property
-    def _forcevec(self):
-        force_1d = np.zeros(self.ordint + 1)
-        return np.kron(np.ones(self.spatialdim), force_1d)
-
-    @cached_property
-    def _dispmat(self):
-        dispmat_1d = np.zeros(self.ordint + 1)
-        dispmat_1d[-1] = 1.0  # Unit diffusion
-        return np.kron(np.eye(self.spatialdim), dispmat_1d).T
-
-    @cached_property
-    def equivalent_discretisation_preconditioned(self):
-        """Discretised IN THE PRECONDITIONED SPACE.
-
-        The preconditioned state transition is the flipped Pascal matrix.
-        The preconditioned process noise covariance is the flipped Hilbert matrix.
-        The shift is always zero.
-
-        Reference: https://arxiv.org/abs/2012.10106
-        """
-
-        state_transition_1d = np.flip(
-            scipy.linalg.pascal(self.ordint + 1, kind="lower", exact=False)
-        )
-        state_transition = np.kron(np.eye(self.spatialdim), state_transition_1d)
-        process_noise_1d = np.flip(scipy.linalg.hilbert(self.ordint + 1))
-        process_noise = np.kron(np.eye(self.spatialdim), process_noise_1d)
-        empty_shift = np.zeros(self.spatialdim * (self.ordint + 1))
-
-        process_noise_cholesky_1d = np.linalg.cholesky(process_noise_1d)
-        process_noise_cholesky = np.kron(
-            np.eye(self.spatialdim), process_noise_cholesky_1d
-        )
-
-        return discrete.DiscreteLTIGaussian(
-            state_trans_mat=state_transition,
-            shift_vec=empty_shift,
-            proc_noise_cov_mat=process_noise,
-            proc_noise_cov_cholesky=process_noise_cholesky,
-            forward_implementation=self.forward_implementation,
-            backward_implementation=self.backward_implementation,
-        )
-
-    def forward_rv(
-        self,
-        rv,
-        t,
-        dt=None,
-        compute_gain=False,
-        _diffusion=1.0,
-        **kwargs,
-    ):
-        if dt is None:
-            raise ValueError(
-                "Continuous-time transitions require a time-increment ``dt``."
-            )
-
-        rv = _apply_precon(self.precon.inverse(dt), rv)
-        rv, info = self.equivalent_discretisation_preconditioned.forward_rv(
-            rv, t, compute_gain=compute_gain, _diffusion=_diffusion
-        )
-
-        info["crosscov"] = self.precon(dt) @ info["crosscov"] @ self.precon(dt).T
-        if "gain" in info:
-            info["gain"] = self.precon(dt) @ info["gain"] @ self.precon.inverse(dt).T
-
-        return _apply_precon(self.precon(dt), rv), info
-
-    def backward_rv(
-        self,
-        rv_obtained,
-        rv,
-        rv_forwarded=None,
-        gain=None,
-        t=None,
-        dt=None,
-        _diffusion=1.0,
-        **kwargs,
-    ):
-        if dt is None:
-            raise ValueError(
-                "Continuous-time transitions require a time-increment ``dt``."
-            )
-
-        rv_obtained = _apply_precon(self.precon.inverse(dt), rv_obtained)
-        rv = _apply_precon(self.precon.inverse(dt), rv)
-        rv_forwarded = (
-            _apply_precon(self.precon.inverse(dt), rv_forwarded)
-            if rv_forwarded is not None
-            else None
-        )
-        gain = (
-            self.precon.inverse(dt) @ gain @ self.precon.inverse(dt).T
-            if gain is not None
-            else None
-        )
-
-        rv, info = self.equivalent_discretisation_preconditioned.backward_rv(
-            rv_obtained=rv_obtained,
-            rv=rv,
-            rv_forwarded=rv_forwarded,
-            gain=gain,
-            t=t,
-            _diffusion=_diffusion,
-        )
-
-        return _apply_precon(self.precon(dt), rv), info
-
-    def discretise(self, dt):
-        """Equivalent discretisation of the process.
-
-        Overwrites matrix-fraction decomposition in the super-class.
-        Only present for user's convenience and to maintain a clean
-        interface. Not used for forward_rv, etc..
-        """
-        state_trans_mat = (
-            self.precon(dt)
-            @ self.equivalent_discretisation_preconditioned.state_trans_mat
-            @ self.precon.inverse(dt)
-        )
-        proc_noise_cov_mat = (
-            self.precon(dt)
-            @ self.equivalent_discretisation_preconditioned.proc_noise_cov_mat
-            @ self.precon(dt).T
-        )
-        zero_shift = np.zeros(len(state_trans_mat))
-
-        # The Cholesky factor of the process noise covariance matrix of the IBM
-        # always exists, even for non-square root implementations.
-        proc_noise_cov_cholesky = (
-            self.precon(dt)
-            @ self.equivalent_discretisation_preconditioned.proc_noise_cov_cholesky
-        )
-
-        return discrete.DiscreteLTIGaussian(
-            state_trans_mat=state_trans_mat,
-            shift_vec=zero_shift,
-            proc_noise_cov_mat=proc_noise_cov_mat,
-            proc_noise_cov_cholesky=proc_noise_cov_cholesky,
-            forward_implementation=self.forward_implementation,
-            backward_implementation=self.forward_implementation,
-        )
-
-
-class IOUP(Integrator, _sde.LTISDE):
-    """Integrated Ornstein-Uhlenbeck process in :math:`d` dimensions."""
-
-    def __init__(
-        self,
-        ordint: int,
-        spatialdim: int,
-        driftspeed: float,
-        forward_implementation="classic",
-        backward_implementation="classic",
-    ):
-        self.driftspeed = driftspeed
-
-        Integrator.__init__(self, ordint=ordint, spatialdim=spatialdim)
-        _sde.LTISDE.__init__(
-            self,
-            driftmat=self._driftmat,
-            forcevec=self._forcevec,
-            dispmat=self._dispmat,
-            forward_implementation=forward_implementation,
-            backward_implementation=backward_implementation,
-        )
-
-    @cached_property
-    def _driftmat(self):
-        driftmat_1d = np.diag(np.ones(self.ordint), 1)
-        driftmat_1d[-1, -1] = -self.driftspeed
-        return np.kron(np.eye(self.spatialdim), driftmat_1d)
-
-    @cached_property
-    def _forcevec(self):
-        force_1d = np.zeros(self.ordint + 1)
-        return np.kron(np.ones(self.spatialdim), force_1d)
-
-    @cached_property
-    def _dispmat(self):
-        dispmat_1d = np.zeros(self.ordint + 1)
-        dispmat_1d[-1] = 1.0  # Unit Diffusion
-        return np.kron(np.eye(self.spatialdim), dispmat_1d).T
-
-    def forward_rv(
-        self,
-        rv,
-        t,
-        dt=None,
-        compute_gain=False,
-        _diffusion=1.0,
-        **kwargs,
-    ):
-        if dt is None:
-            raise ValueError(
-                "Continuous-time transitions require a time-increment ``dt``."
-            )
-
-        # Fetch things into preconditioned space
-        rv = _apply_precon(self.precon.inverse(dt), rv)
-
-        # Apply preconditioning to system matrices
-        self.driftmat = self.precon.inverse(dt) @ self.driftmat @ self.precon(dt)
-        self.forcevec = self.precon.inverse(dt) @ self.forcevec
-        self.dispmat = self.precon.inverse(dt) @ self.dispmat
-
-        # Discretise and propagate
-        discretised_model = self.discretise(dt=dt)
-        rv, info = discretised_model.forward_rv(
-            rv, t, compute_gain=compute_gain, _diffusion=_diffusion
-        )
-
-        # Undo preconditioning and return
-        rv = _apply_precon(self.precon(dt), rv)
-        info["crosscov"] = self.precon(dt) @ info["crosscov"] @ self.precon(dt).T
-        if "gain" in info:
-            info["gain"] = self.precon(dt) @ info["gain"] @ self.precon.inverse(dt).T
-
-        self.driftmat = self.precon(dt) @ self.driftmat @ self.precon.inverse(dt)
-        self.forcevec = self.precon(dt) @ self.forcevec
-        self.dispmat = self.precon(dt) @ self.dispmat
-
-        return rv, info
-
-    def backward_rv(
-        self,
-        rv_obtained,
-        rv,
-        rv_forwarded=None,
-        gain=None,
-        t=None,
-        dt=None,
-        _diffusion=1.0,
-        **kwargs,
-    ):
-        if dt is None:
-            raise ValueError(
-                "Continuous-time transitions require a time-increment ``dt``."
-            )
-
-        # Fetch things into preconditioned space
-        rv_obtained = _apply_precon(self.precon.inverse(dt), rv_obtained)
-        rv = _apply_precon(self.precon.inverse(dt), rv)
-        rv_forwarded = (
-            _apply_precon(self.precon.inverse(dt), rv_forwarded)
-            if rv_forwarded is not None
-            else None
-        )
-        gain = (
-            self.precon.inverse(dt) @ gain @ self.precon.inverse(dt).T
-            if gain is not None
-            else None
-        )
-
-        # Apply preconditioning to system matrices
-        self.driftmat = self.precon.inverse(dt) @ self.driftmat @ self.precon(dt)
-        self.forcevec = self.precon.inverse(dt) @ self.forcevec
-        self.dispmat = self.precon.inverse(dt) @ self.dispmat
-
-        # Discretise and propagate
-        discretised_model = self.discretise(dt=dt)
-        rv, info = discretised_model.backward_rv(
-            rv_obtained=rv_obtained,
-            rv=rv,
-            rv_forwarded=rv_forwarded,
-            gain=gain,
-            t=t,
-            _diffusion=_diffusion,
-        )
-
-        # Undo preconditioning and return
-        rv = _apply_precon(self.precon(dt), rv)
-        self.driftmat = self.precon(dt) @ self.driftmat @ self.precon.inverse(dt)
-        self.forcevec = self.precon(dt) @ self.forcevec
-        self.dispmat = self.precon(dt) @ self.dispmat
-        return rv, info
-
-
-class Matern(Integrator, _sde.LTISDE):
-    """Matern process in :math:`d` dimensions."""
-
-    def __init__(
-        self,
-        ordint: int,
-        spatialdim: int,
-        lengthscale: float,
-        forward_implementation="classic",
-        backward_implementation="classic",
-    ):
-
-        self.lengthscale = lengthscale
-
-        Integrator.__init__(self, ordint=ordint, spatialdim=spatialdim)
-        _sde.LTISDE.__init__(
-            self,
-            driftmat=self._driftmat,
-            forcevec=self._forcevec,
-            dispmat=self._dispmat,
-            forward_implementation=forward_implementation,
-            backward_implementation=backward_implementation,
-        )
-
-    @property
-    def _driftmat(self):
-        driftmat = np.diag(np.ones(self.ordint), 1)
-        nu = self.ordint + 0.5
-        D, lam = self.ordint + 1, np.sqrt(2 * nu) / self.lengthscale
-        driftmat[-1, :] = np.array(
-            [-scipy.special.binom(D, i) * lam ** (D - i) for i in range(D)]
-        )
-        return np.kron(np.eye(self.spatialdim), driftmat)
-
-    @property
-    def _forcevec(self):
-        force_1d = np.zeros(self.ordint + 1)
-        return np.kron(np.ones(self.spatialdim), force_1d)
-
-    @property
-    def _dispmat(self):
-        dispmat_1d = np.zeros(self.ordint + 1)
-        dispmat_1d[-1] = 1.0  # Unit diffusion
-        return np.kron(np.eye(self.spatialdim), dispmat_1d).T
-
-    def forward_rv(
-        self,
-        rv,
-        t,
-        dt=None,
-        compute_gain=False,
-        _diffusion=1.0,
-        **kwargs,
-    ):
-        if dt is None:
-            raise ValueError(
-                "Continuous-time transitions require a time-increment ``dt``."
-            )
-
-        # Fetch things into preconditioned space
-        rv = _apply_precon(self.precon.inverse(dt), rv)
-
-        # Apply preconditioning to system matrices
-        self.driftmat = self.precon.inverse(dt) @ self.driftmat @ self.precon(dt)
-        self.forcevec = self.precon.inverse(dt) @ self.forcevec
-        self.dispmat = self.precon.inverse(dt) @ self.dispmat
-
-        # Discretise and propagate
-        discretised_model = self.discretise(dt=dt)
-        rv, info = discretised_model.forward_rv(
-            rv, t, compute_gain=compute_gain, _diffusion=_diffusion
-        )
-
-        # Undo preconditioning and return
-        rv = _apply_precon(self.precon(dt), rv)
-        info["crosscov"] = self.precon(dt) @ info["crosscov"] @ self.precon(dt).T
-        if "gain" in info:
-            info["gain"] = self.precon(dt) @ info["gain"] @ self.precon.inverse(dt).T
-
-        self.driftmat = self.precon(dt) @ self.driftmat @ self.precon.inverse(dt)
-        self.forcevec = self.precon(dt) @ self.forcevec
-        self.dispmat = self.precon(dt) @ self.dispmat
-
-        return rv, info
-
-    def backward_rv(
-        self,
-        rv_obtained,
-        rv,
-        rv_forwarded=None,
-        gain=None,
-        t=None,
-        dt=None,
-        _diffusion=1.0,
-        **kwargs,
-    ):
-        if dt is None:
-            raise ValueError(
-                "Continuous-time transitions require a time-increment ``dt``."
-            )
-
-        # Fetch things into preconditioned space
-        rv_obtained = _apply_precon(self.precon.inverse(dt), rv_obtained)
-        rv = _apply_precon(self.precon.inverse(dt), rv)
-        rv_forwarded = (
-            _apply_precon(self.precon.inverse(dt), rv_forwarded)
-            if rv_forwarded is not None
-            else None
-        )
-        gain = (
-            self.precon.inverse(dt) @ gain @ self.precon.inverse(dt).T
-            if gain is not None
-            else None
-        )
-
-        # Apply preconditioning to system matrices
-        self.driftmat = self.precon.inverse(dt) @ self.driftmat @ self.precon(dt)
-        self.forcevec = self.precon.inverse(dt) @ self.forcevec
-        self.dispmat = self.precon.inverse(dt) @ self.dispmat
-
-        # Discretise and propagate
-        discretised_model = self.discretise(dt=dt)
-        rv, info = discretised_model.backward_rv(
-            rv_obtained=rv_obtained,
-            rv=rv,
-            rv_forwarded=rv_forwarded,
-            gain=gain,
-            t=t,
-            _diffusion=_diffusion,
-        )
-
-        # Undo preconditioning and return
-        rv = _apply_precon(self.precon(dt), rv)
-        self.driftmat = self.precon(dt) @ self.driftmat @ self.precon.inverse(dt)
-        self.forcevec = self.precon(dt) @ self.forcevec
-        self.dispmat = self.precon(dt) @ self.dispmat
-        return rv, info
-
-
-def _apply_precon(precon, rv):
-
-    # There is no way of checking whether `rv` has its Cholesky factor computed already or not.
-    # Therefore, since we need to update the Cholesky factor for square-root filtering,
-    # we also update the Cholesky factor for non-square-root algorithms here,
-    # which implies additional cost.
-    # See Issues #319 and #329.
-    # When they are resolved, this function here will hopefully be superfluous.
-
-    new_mean = precon @ rv.mean
-    new_cov_cholesky = precon @ rv.cov_cholesky  # precon is diagonal, so this is valid
-    new_cov = new_cov_cholesky @ new_cov_cholesky.T
-
-    return randvars.Normal(new_mean, new_cov, cov_cholesky=new_cov_cholesky)
